make execute_circuit also output a dict of Counter
