+++ src/qibocal/protocols/tomographies/two_qubit_state_tomography.py
@@ -1,6 +1,5 @@
 import json
 from collections import Counter, defaultdict
-from collections.abc import Sequence
 from copy import deepcopy
 from dataclasses import dataclass, field
 from itertools import product
@@ -28,6 +27,7 @@ from qibocal.auto.transpile import (
     get_compiler,
 )
 from qibocal.calibration import CalibrationPlatform
+from qibocal.protocols.utils import marginalize_qubit_counts
 
 from ..utils import table_dict, table_html
 from .state_tomography import StateTomographyParameters, plot_reconstruction
@@ -98,25 +98,6 @@ class StateTomographyResults(Results):
     """State fidelity."""
 
 
-def marginalize_qubit_counts(counts: Counter, indices: Sequence[int] | int):
-    """
-    Extract marginal distribution from measurement counts over selected qubit indices.
-
-    Args:
-        counts: Counter mapping bitstrings to counts (e.g., {'0101': 10, ...})
-        indices: Qubit positions to marginalize over.
-
-    Returns:
-        Counter of the marginal distribution.
-    """
-    out = Counter()
-    indices_list = [indices] if isinstance(indices, int) else indices
-    for state, count in counts.items():
-        reduced = "".join(state[i] for i in indices_list)
-        out[reduced] += count
-    return out
-
-
 def _acquisition(
     params: StateTomographyParameters,
     platform: CalibrationPlatform,
@@ -173,7 +154,7 @@ def _acquisition(
         )
 
         for i, pair in enumerate(targets):
-            frequencies = marginalize_qubit_counts(results, (2 * i, 2 * i + 1))
+            frequencies = marginalize_qubit_counts(results[0], (2 * i, 2 * i + 1))
             simulation_probabilities = simulation_result.probabilities(
                 qubits=(2 * i, 2 * i + 1)
             )
